# Add `MessagePackSerializer.MaxAsyncBuffer` to speed up small async deserializations