--- a/src/Nerdbank.MessagePack/netstandard2.0/PublicAPI.Unshipped.txt
+++ b/src/Nerdbank.MessagePack/netstandard2.0/PublicAPI.Unshipped.txt
@@ -195,6 +195,8 @@ Nerdbank.MessagePack.MessagePackSerializer.InternStrings.get -> bool
 Nerdbank.MessagePack.MessagePackSerializer.InternStrings.init -> void
 Nerdbank.MessagePack.MessagePackSerializer.LibraryExtensionTypeCodes.get -> Nerdbank.MessagePack.LibraryReservedMessagePackExtensionTypeCode!
 Nerdbank.MessagePack.MessagePackSerializer.LibraryExtensionTypeCodes.init -> void
+Nerdbank.MessagePack.MessagePackSerializer.MaxAsyncBuffer.get -> int
+Nerdbank.MessagePack.MessagePackSerializer.MaxAsyncBuffer.init -> void
 Nerdbank.MessagePack.MessagePackSerializer.PreserveReferences.get -> bool
 Nerdbank.MessagePack.MessagePackSerializer.PreserveReferences.init -> void
 Nerdbank.MessagePack.MessagePackSerializer.PropertyNamingPolicy.get -> Nerdbank.MessagePack.MessagePackNamingPolicy?