Since synchronous deserialization is substantially faster than async deserialization, we prefer sync. But sync requires that all msgpack data be pre-buffered. That's a reasonable trade-off, assuming the msgpack data is reasonably small. When it is large, the slower async path may be preferable to avoid unbounded memory consumption.

Closes #155

## Before

| Method            | Mean     | Error    | StdDev  | Ratio | RatioSD | Gen0    | Gen1   | Allocated | Alloc Ratio |
|------------------ |---------:|---------:|--------:|------:|--------:|--------:|-------:|----------:|------------:|
| Sync_Deserialize  | 111.4 us | 19.72 us | 1.08 us |  1.00 |    0.01 | 11.4746 | 1.9531 |  70.34 KB |        1.00 |
| Async_Deserialize | 327.0 us |  6.20 us | 0.34 us |  2.94 |    0.02 | 11.2305 | 1.9531 |  70.49 KB |        1.00 |

## After


Method | Categories | Mean | Error | StdDev | Ratio | Gen0 | Gen1 | Allocated | Alloc Ratio
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
Sync_Deserialize | map-init,Deserialize | 114.37 μs | 10.251 μs | 0.562 μs | 1.00 | 11.4746 | 1.9531 | 72024 B | 1.00
Async_Deserialize | map-init,Deserialize | 104.00 μs | 2.420 μs | 0.133 μs | 0.91 | 11.4746 | 1.9531 | 72080 B | 1.00

Not all benchmarks show 0 tax for async deserialization for small jobs. The above works because it's a reasonably _large_ array but still under the threshold for synchronous deserialization. But very tiny deserializations still pay a tax due to buffering overhead (even if it's already in memory) that sync deserialization APIs can avoid.